statsd: aggregate metrics from a chunk in batch
Process all metrics in a chunk together.

This is necessary for custom Prometheus-like histogram metrics. For
example, when creating an `http_response_duration_seconds` metric with
buckets 1...100 and a flush interval in Telegraf set to 30 seconds,
the old version of the code would process each sample separately. If a
flush interval occurred, only the processed metrics would be flushed
with the current timestamp. The next metrics would be flushed in the
following interval with a different timestamp (+30s), causing issues
with histogram calculation. In the current version, all samples are
processed together.